+++ packages/kernel/src/application/facades/index.ts
@@ -1,1 +1,2 @@
 export * from "./meta-kernel.facade";
+export * from "./migration-safety.facade";