+++ packages/runtime/src/application/facades/index.ts
@@ -1,0 +1,2 @@
+export * from "./runtime-view.facade";
+export * from "./runtime-interaction.facade";