+++ packages/runtime/src/application/index.ts
@@ -1,2 +1,1 @@
-export * from "./facades/runtime-view.facade";
-export * from "./facades/runtime-interaction.facade";
+export * from "./facades";