+++ packages/runtime/src/index.ts
@@ -1,6 +1,2 @@
 export * from "./core";
-export { executeRuntimeGatewayView, executeRuntimeView } from "./application/facades/runtime-view.facade";
-export {
-  createRecordingRuntimeInteractionPort,
-  executeRuntimeInteractionPlan
-} from "./application/facades/runtime-interaction.facade";
+export * from "./application/facades";